requirements_local.txt
altair==5.5.0
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==24.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.7.0
easyocr==1.7.2
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==0.8.0
google-generativeai==0.8.3
google_search_results==2.4.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gTTS==2.5.4
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lxml==5.3.1
mailchimp-transactional==1.0.56
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.23.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playsound==1.2.2
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.41.1
svglib==1.5.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.1.10
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
watchdog==6.0.0
weasyprint==64.0
webencodings==0.5.1
websockets==14.2
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
